fix(sql): idempotent pooled-connection close; test placement and doc clarifications
- Add _closed guard to _PooledConnection.close() so double-close in the
  execute-retry path (explicit close + outer finally) is a no-op; the
  underlying TDS socket now receives close() exactly once.
- Assert bad_conn.close.assert_called_once() in the execute-retry test to
  lock in the idempotency contract.
- Move TestD22InvalidItemType from test_sql.py to test_resolver.py where
  it belongs (tests resolver.py, not sql.py); remove now-unused
  `from pathlib import Path` import from test_sql.py.
- Expand run_query docstring to state execute-phase retry fires at most
  once regardless of SQL_TRANSIENT_MAX_RETRIES.
- Add comment at _NATIVE_ERROR_RE explaining the intentional distinction
  between Error:/error (alt-1) and Error\b (alt-2) and that
  first-match-wins is intentional.
