+++ tests/unit/test_sql.py
@@ -4,7 +4,6 @@ from __future__ import annotations
 
 import threading
 from contextlib import closing
-from pathlib import Path
 from unittest.mock import MagicMock
 
 import pytest
@@ -1555,6 +1554,10 @@ class TestD10RetryBoundary:
 
         assert rows == [(42,)]
         assert mock_mssql.connect.call_count == 2
+        # _PooledConnection.close() is idempotent; despite being called once
+        # explicitly (retry path) and once in the outer finally, the underlying
+        # bad_conn must receive close() exactly once.
+        bad_conn.close.assert_called_once()
 
     def test_fetch_one_retried_once_on_transient_execute_error(
         self, monkeypatch: pytest.MonkeyPatch
@@ -1580,38 +1583,6 @@ class TestD10RetryBoundary:
 
         assert rows == [(99,)]
         assert mock_mssql.connect.call_count == 2
-
-
-# ---------------------------------------------------------------------------
-# D22 -- reject invalid item_type with a clear FabricError
-# ---------------------------------------------------------------------------
-
-
-class TestD22InvalidItemType:
-    """D22: resolver.item() must raise FabricError immediately for unknown item_type."""
-
-    @pytest.mark.asyncio
-    async def test_invalid_item_type_raises_fabric_error_name_path(self, tmp_path: Path) -> None:
-        """Passing an unknown item_type by name must raise FabricError before any API call."""
-        from unittest.mock import AsyncMock  # noqa: PLC0415
-
-        from fabric_dw.cache import LookupCache  # noqa: PLC0415
-        from fabric_dw.exceptions import FabricError  # noqa: PLC0415
-        from fabric_dw.resolver import Resolver  # noqa: PLC0415
-
-        cache = LookupCache(path=tmp_path / "lookup.json")
-        # Use a minimal mock for the HTTP client -- it must NOT be called.
-        mock_http = MagicMock()
-        mock_http.iter_paginated = AsyncMock()
-        resolver = Resolver(http=mock_http, cache=cache)
-
-        ws_guid = "00000000-0000-0000-0000-000000000001"
-
-        with pytest.raises(FabricError, match="Unknown item_type"):
-            await resolver.item(ws_guid, "MyWarehouse", item_type="NotARealType")
-
-        # The HTTP layer must NOT have been called -- error raised before pagination.
-        mock_http.iter_paginated.assert_not_called()
 
 
 # ---------------------------------------------------------------------------
